Rename example service
