--- src/services/example-service/index.ts
@@ -1,3 +1,0 @@
-export * from './client';
-export * from './getExample';
-export * from './types';
