+++ src/services/exampleService/getExample.ts
@@ -1,5 +1,5 @@
-import { makeRequest } from 'src/services/example-service/client';
-import { Example, ResponsePromise } from 'src/services/example-service/types';
+import { makeRequest } from 'src/services/exampleService/client';
+import { Example, ResponsePromise } from 'src/services/exampleService/types';
 
 
 export const getExample = async(id: string): ResponsePromise<Example> => (
