+++ src/services/exampleService/index.ts
@@ -1,0 +1,3 @@
+export * from 'src/services/exampleService/client';
+export * from 'src/services/exampleService/getExample';
+export * from 'src/services/exampleService/types';
