+++ src/services/exampleService/types.ts
